+++ client/app/routes/reviews-controller.js
@@ -17,19 +17,22 @@ angular.module("reviews").controller("ReviewsController", ['$scope', '$http', fu
         };
 
         this.delRev = function (revid, docid) { // deletes review to the document depending on the actual context
-            $http.delete('//aleph.inesc-id.pt/vre/api/reviews/' + $scope.type + '/' + docid + '/' + revid)
-                .success(function () {
-                    $http.delete('//aleph.inesc-id.pt/vre/api/delete/reviews/' + revid)
-                        .success(function () {
-                            location.reload();
-                        })
-                        .error(function () {
-                            console.log("Error: Could not remove review from reviews");
-                        });
-                })
-                .error(function () {
-                    console.log("Error: Could not remove review from document");
-                });
+            var r = confirm("This will remove the review permanently.");
+            if (r == true) {
+                $http.delete('//aleph.inesc-id.pt/vre/api/reviews/' + $scope.type + '/' + docid + '/' + revid)
+                    .success(function () {
+                        $http.delete('//aleph.inesc-id.pt/vre/api/delete/reviews/' + revid)
+                            .success(function () {
+                                location.reload();
+                            })
+                            .error(function () {
+                                console.log("Error: Could not remove review from reviews");
+                            });
+                    })
+                    .error(function () {
+                        console.log("Error: Could not remove review from document");
+                    });
+            }
         };
     }]
 );
